+++ src/inspectors/mod.rs
@@ -7,7 +7,9 @@ pub use external_functions::{
 pub use function_source::{
     FunctionSourceInspector, FunctionSourceOutput, ResolvedSymbol, SymbolKind,
 };
+pub use storage_layout::{StorageLayoutInspector, StorageLayoutOutput, StorageVariableInfo};
 
 mod compile;
 mod external_functions;
 mod function_source;
+mod storage_layout;